requirements.txt
Django==1.9.6
django-cors-headers==1.1.0
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
PyJWT==1.4.2
psycopg2==2.6.2
